Time: 21 ms (64.69%) | Memory: 70.7 MB (14.30%) - LeetSync

+++ 1605-minimum-number-of-days-to-make-m-bouquets/minimum-number-of-days-to-make-m-bouquets.cpp
@@ -19,7 +19,7 @@ public:
         return m == 0;
     }
     int minDays(vector<int>& bloomDay, int m, int k) {
-        double Product = (double)m * (double)k;
+        long long int Product = (long long )m * (long long)k;
         if(bloomDay.size() < Product)
         return -1;
         int start = *min_element(bloomDay.begin(), bloomDay.end());
